Makes naming schemes of data parsing functions uniform
