+++ sndata/utils/data_parsing.py
@@ -11,9 +11,7 @@ from typing import Union
 
 import numpy as np
 import sncosmo
-from astropy.table import Table
 
-from . import unit_conversion
 from ..exceptions import NoDownloadedData
 
 
@@ -97,100 +95,6 @@ def parse_vizier_table_descriptions(readme_path: Union[Path, str]):
             table_descriptions[table_num] = table_desc
 
     return table_descriptions
-
-
-def parse_snoopy_path(path: str):
-    """Return data from a snoopy file as an astropy table
-
-    Args:
-        path: The file path of a snoopy input file
-
-    Returns:
-        An astropy table with columns 'time', 'band', 'mag', and 'mag_err'
-    """
-
-    out_table = Table(
-        names=['time', 'band', 'mag', 'mag_err'],
-        dtype=[float, object, float, float]
-    )
-
-    with open(path) as ofile:
-        # Get meta data from first line
-        name, z, ra, dec = ofile.readline().split()
-        out_table.meta['obj_id'] = name
-        out_table.meta['ra'] = float(ra)
-        out_table.meta['dec'] = float(dec)
-        out_table.meta['z'] = float(z)
-        out_table.meta['z_err'] = None
-
-        # Read photometric data from the rest of the file
-        band = None
-        for line in ofile.readlines():
-            line_list = line.split()
-            if line.startswith('filter'):
-                band = line_list[1]
-                continue
-
-            time, mag, mag_err = line_list
-            out_table.add_row([time, band, mag, mag_err])
-
-    out_table['time'] = unit_conversion.convert_to_jd(out_table['time'], format='snpy')
-    return out_table
-
-
-def read_csp_spectroscopy_file(path: str, format_table: bool = False) -> Table:
-    """Read a file path of spectroscopic data from CSP
-
-    Args:
-        path: Path of file to read
-        format_table: Format table to a sndata standard format
-
-    Returns:
-        An astropy table with file data and meta data
-    """
-
-    path = Path(path)
-    obj_id = '20' + path.name.split('_')[0].lstrip('SN')
-
-    # Handle the single file with a different data model:
-    # This file has three columns instead of two
-    if path.stem == 'SN07bc_070409_b01_BAA_IM':
-        data = Table.read(path, format='ascii', names=['wavelength', 'flux', '_'])
-        data.remove_column('_')
-
-    else:
-        data = Table.read(path, format='ascii', names=['wavelength', 'flux'])
-
-    # Read the table meta data
-    file_comments = data.meta['comments']
-    redshift = float(file_comments[1].lstrip('Redshift: '))
-    obs_date = float(file_comments[3].lstrip('JDate_of_observation: '))
-    epoch = float(file_comments[4].lstrip('Epoch: '))
-
-    # Add meta data to output table according to sndata standard
-    data.meta['obj_id'] = obj_id
-    data.meta['ra'] = None
-    data.meta['dec'] = None
-    data.meta['z'] = redshift
-    data.meta['z_err'] = None
-    del data.meta['comments']
-
-    data['time'] = obs_date
-    if format_table:
-        # Add remaining columns. These values are constant for a single file
-        # (i.e. a single spectrum) but vary across files (across spectra)
-        _, _, w_range, telescope, instrument = path.stem.split('_')
-        data['epoch'] = epoch
-        data['wavelength_range'] = w_range
-        data['telescope'] = telescope
-        data['instrument'] = instrument
-
-        # Enforce an intuitive column order
-        data = data[[
-            'time', 'wavelength', 'flux', 'epoch',
-            'wavelength_range', 'telescope', 'instrument']]
-
-    return data
 
 
 def register_filter_file(file_path: str, filter_name: str, force: bool = False):
